+++ Arma3ObjectBuilder/logger.py
