+++ Arma3ObjectBuilder/mcfg/importer.py
@@ -8,7 +8,7 @@ import subprocess
 
 from .. import get_prefs
 from .. import config
-from ..utilities.logger import ProcessLogger
+from ..logger import ProcessLogger
 
 
 class Bone():
